+++ .github/workflows/flights_test.yml
@@ -52,10 +52,8 @@ jobs:
           DATABRICKS_HOST: ${{ secrets.DATABRICKS_HOST }}
           DATABRICKS_TOKEN: ${{ secrets.DATABRICKS_TOKEN_TST }}
           DATABRICKS_CLUSTER_ID: ${{ secrets.DATABRICKS_CLUSTER_ID }}
-          # SPARK_REMOTE: ${{secrets.SPARK_REMOTE_TST}}   
         run: |
           pytest tests/ 
-      #-k test_get_flight_schema__valid
       
       - uses: databricks/setup-cli@main
 
@@ -81,10 +79,19 @@ jobs:
       - uses: databricks/setup-cli@main
 
       - shell: bash
-        name: Run validation job
+        name: Run notebook validation
         run: |
           set -o pipefail
-          databricks bundle run -t test notebook_validation_job --refresh-all 2>&1 | tee output.log
+          databricks bundle run notebook_validation_job -t test --refresh-all 2>&1 | tee output.log
         env:
           DATABRICKS_TOKEN: ${{ secrets.DATABRICKS_TOKEN_TST }}
           DATABRICKS_BUNDLE_ENV: tst
+
+      # - shell: bash
+      #   name: Run pipeline validation
+      #   run: |
+      #     set -o pipefail
+      #     databricks bundle run flights_simple_dlt_validation -t test
+      #   env:
+      #     DATABRICKS_TOKEN: ${{ secrets.DATABRICKS_TOKEN_TST }}
+      #     DATABRICKS_BUNDLE_ENV: tst
