+++ databricks.yml
@@ -7,6 +7,13 @@ bundle:
 
 include:
   - resources/*.yml
+  - resources/dlt/*.yml
+  
+## Uncomment below to build wheel as part of bundle
+#artifacts:
+#  default:
+#    type: whl
+#    path: .
 
 variables:
   catalog:
@@ -15,6 +22,10 @@ variables:
     default: "dustinvannoy_dev"
   all_purpose_cluster_id:
     default: "0326-164201-qhvj1pwh"
+  flights_test_schema:
+    default: "flights_test"
+  flights_dlt_schema:
+    default: "dustinvannoy_dev"
 
 targets:
   # The 'dev' target, used for development purposes.
@@ -43,7 +54,8 @@ targets:
       user_name: ${workspace.current_user.userName}
     variables:
       all_purpose_cluster_id: "0725-142209-7jvag949"
-    
+      flights_test_schema: flights_validation_test
+      flights_dlt_schema: dustinvannoy_test
 
   staging:
     # For staging deployments, we only have a single copy, so we override the
@@ -58,6 +70,8 @@ targets:
       user_name: ${workspace.current_user.userName}
     variables:
       all_purpose_cluster_id: "0326-164201-qhvj1pwh"
+      flights_test_schema: flights_validation_staging
+      flights_dlt_schema: dustinvannoy_staging
 
   # The 'prod' target, used for production deployment.
   prod:
@@ -75,3 +89,5 @@ targets:
       user_name: ${workspace.current_user.userName}
     variables:
       all_purpose_cluster_id: "0326-164201-qhvj1pwh"
+      flights_test_schema: flights_validation_prod
+      flights_dlt_schema: dustinvannoy_prod
