+++ setup.py
@@ -7,7 +7,7 @@ Source code is in ./src/. Run 'python setup.py sdist bdist_wheel' to build.
 from setuptools import setup, find_packages
 
 import sys
-sys.path.append('./flights')
+sys.path.append('.')
 
 import flights
 
@@ -18,7 +18,7 @@ setup(
     author="training@dustinvannoy.com",
     description="Flights package",
     packages=find_packages(where='.'),
-    package_dir={'': 'flights'},
+    package_dir={'': '.'},
     # entry_points={"entry_points": "main=flights.main:main"},
     install_requires=["setuptools"],
 )
